move MapLibreMapFactory
